ENH: Add a frozen column for the plating UI
This adds a frozen column to the UI, since SlickGrid doesn't support
frozen columns, this was implemented by having two grids, one with a
single column (row names) and the other with the contents of the plate.

Note that the height of the containers is dynamically adjusted
(depending on the number of rows) so that we don't have to handle
scrolling position synchronization between the two grids.

Fixes #151