ENH: Add discard plate functionality
This patch makes two additions to the Plate class, first it adds a
"discarded" setter, and second it extends the functionality in the
list_plates method to be able to retrieve (or not) plates that have been
marked as discarded.

The PATCH handler is now aware about the discarded property. The UI uses
this route to make changes from the plate list.

Fixes #150