BUG: Fix broken copy-paste
The assignment of the `this.blankNames` property was being done
asynchronously on a per-cell basis. This worked fine if the user inputed
a value by hand, as it gave enough time for the request to return.
However, when someone copy-pasted text into the grid, the request
wouldn't return in time and the `this.blankNames` property wouldn't be
correctly assigned leading to a ReferenceError.

This is now fixed by getting the contents of `this.blankNames` before
initializing the whole grid. This way every cell has access to the data
before instantiation.

Fixes #243