Fix Orig_name mod affecting sample_name column
As orig_name also is used to generate the sample_name column output,
changes to orig_name to drop the prepended Study ID were also impacting
the prepended Study ID in the sample_name column. The proposed solution
for the near term is to have a second orig_name2 column, that becomes
the final Orig_name column in the output, while the original orig_name
column is used to generate the sample_name column before being deleted
from the final output.