WIP Gen shotgun prep files (#502)
* Shotgun prep file generation

Work in Progress

* Documentation/flake8 pass

* An initial test file for prep file output testing

* Addressing change requests made in meeting

* Migration from CSV module to DataFrame

Partial implementation of munging using DataFrame.

* All columns are present and filled out.

All columns are present and filled out. There are some partial null
columns where rows represent control data; this will be resolved
in a later push.

* All columns completed

* Remove debugging code

* Fixed testing

* Added constraints to labman.well row and col columns

* First set of changes in response to feedback

* Additional requested changes

Added additional requested changes, as well as a few statements to write
both raw and final output to disk for debugging purposes. Will remove
from final version. Note that this comment in this issue needs to be
resolved to everyone's satisfaction:

https://github.com/jdereus/labman/pull/502#discussion_r282764816

* Changes requested, Debugging removed