Modified PlateViewer colors + Legend
Text for invalid specimen ids implied id would not exist across all
studies, when it's only against the studies associated with the plate.

Added a new yellow candy-cane/barber-bar to represent a cell that was
populated using a multiselect paste and has more than one possible
specimen ID match across the attached studies. This is so the user can
be alerted, click on the cell, and select from the drop-down that should
appear.

Modified the colors to be more appealing. Looking for collective
feedback. Bold red is gone, and an invalid specimen ID is now the red
candy-cane color. Blue candy-cane color is the same as before,
representing a specimen ID that has already been plated on another
plate.

A new Yellow/Orange candy-cane stripe identifies values that match more
than one possible specimen ID from the attached studies.

The candy-cane types are mutually exclusive; a cell can't be both valid
and invalid (0 possible matches), invalid and matching 2 or more
possible specimen ids, etc.

However, any of the above types can still have duplicate entries on the
current plate. Hence, the platemapper now identifies duplicate entries
by wrapping the cell in a purple border. This way, there doesn't have to
be a hierarchy defining which error should be presented over another,
and it won't be random based on implementation.

Lastly, as the purple border can overwrite the green border for a cell
with comment, cells with comments will now have green lettering, rather
than a green border.