BUG: Selectively register CECM plugin on PV init
This checks to see if the multiSelectCheckbox is or isn't checked
(it defaults to being checked, but it can be unchecked if the user
uses the back/fwd buttons of their browser to return to this page
after unchecking it).

This is the second of the "warts" remaining in my solution of #354,
so my fork should now be ready for a PR.

Also, I'm gonna tag #562 in this -- this sort of solution might
work there (although just using $(document).ready to force the <select>
to start on `Choose plate type...` would probably be a better option).