BUG: fix another instance of #562 in plating page
As discussed with @charles-cowart, I suspect this problem occurs
throughout multiple other pages in LabControl (e.g. it's easy to
replicate on the "Create working primer plates" admin page -- populate
all of the fields with a bunch of values until the "Create" button is
enabled, press the back button, then press the forward button, and
all of those fields will still be populated but the "Create" button
will be disabled).

The easiest general-case solution is just saying "please don't use the
back or forward buttons", but IDK if that's really tenable for this
project. For making this sorta stuff work with these browser buttons,
more code needs to be written throughout the affected pages that
resets form values/etc. -- and that'll take likely a nontrivial amount
of dev time/testing.