TST: Explicitly test & improve study samples limit
Previously in this PR, this wasn't directly tested, and things were
kind of funky -- Study.samples() said that it expected limit to be an
int, but the limit obtained from get_argument() in
StudySamplesHandler.get() was actually a str.

Major changes:

- Study.samples() now explicitly requires that limit is already an
  int.

- Added test_samples_with_limit() to TestStudy accordingly.

- StudySamplesHandler.get() now converts limit to an int (if it's been
  specified), before calling Study.samples(). If int() fails, then
  it'll raise a ValueError. (Similarly, Study.samples() raises a
  ValueError for invalid limits.) If StudySamplesHandler.get() runs
  into a ValueError, it sets a status code of 400 ("Bad Request").

- Added various invalid limit tests to
  test_get_study_samples_handler() accordingly.

Running tests on my system seems to explode the database on my system
(I'll look into getting this fixed soon -- since I've been working on
the front-end mostly this hasn't been a problem), so I'm going to just
kick this up to Travis for the time being.