TST: Fix error in exp output + beef up test cases
For some reason, I forgot that there were 9 "SKM" samples in the
test dataset. The thing I was trying to test with the

response = self.get('/study/1/samples?term=SKM&limit=30')

test was that the property of limits not doing anything if greater
than the normal amount of output still worked, even if a term was
specified. This test should be fixed now.